project 2: quiz/requirements.txt
ollama==0.4.4
pinecone==5.4.2
python-dotenv==1.0.1
sentence_transformers==3.3.1
streamlit==1.40.2
youtube_transcript_api==0.6.3
yt_dlp==2024.12.13
